delay creation of xml reader
