+++ build.gradle
@@ -35,6 +35,7 @@ dependencies {
     implementation 'commons-io:commons-io:2.11.0'
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'org.nanohttpd:nanohttpd:2.3.1'
+    testImplementation 'org.mockito:mockito-core:4.4.0'
     testRuntimeOnly 'org.slf4j:slf4j-simple:1.7.36'
     testRuntimeOnly 'org.apache.logging.log4j:log4j-to-slf4j:2.17.2'
 }
